psycherank/requirements.txt
asgiref==3.5.2
Django==4.1.1
django-cors-headers==3.13.0
django-jsonfield==1.4.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
psycopg2==2.9.3
Pygments==2.13.0
PyJWT==2.5.0
python-decouple==3.6
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.2
